refactor: move NonceMap into utils
